test: API 6. 학습지 학습 통계 기능 관련 테스트 코드 작성 - 정답률 계산, 학습지 통계 데이터 생성 기능

+++ src/main/kotlin/org/freewheelin/homeschoolmaterials/domain/problem/AnswerRateAnalyzer.kt
@@ -9,7 +9,7 @@ class AnswerRateAnalyzer {
 		val totalCount = submittedProblems.size
 		val answeredCount = submittedProblems.filter { it.isAnswered }.size
 
-		return ((answeredCount / totalCount) * 100) * 10.0 / 10.0
+		return ((answeredCount.toDouble() / totalCount.toDouble()) * 100) * 10.0 / 10.0
 	}
 
 	fun calculateProblemAnswerRate(problemId: Long, submittedProblems: List<SubmittedProblemDto>): Double {
@@ -17,6 +17,6 @@ class AnswerRateAnalyzer {
 		val problemCount = filteredProblems.size
 		val answeredCount = filteredProblems.filter { it.isAnswered }.size
 
-		return ((answeredCount / problemCount) * 100) * 10.0 / 10.0
+		return ((answeredCount.toDouble() / problemCount.toDouble()) * 100) * 10.0 / 10.0
 	}
 }
